Finsihed Logic for delete card form and modal.

+++ src/Components/DeleteCardForm/DeleteCardForm.jsx
@@ -2,13 +2,17 @@ import { useState } from "react";
 import axios from "axios";
 import "./AddCardForm.css";
 
-const DeleteCardForm = ({ getCardsforCollection, setDeleteShowModal }) => {
-  const [collection, setCollection] = useState(0);
-  const [CardId, setCardId] = useState(0);
+const DeleteCardForm = ({
+  getCardsforCollection,
+  setDeleteShowModal,
+  activeCollection,
+  displayCard,
+}) => {
+  const [validationAnswer, setValidationAnswer] = useState("");
 
   async function deleteCard() {
     let response = await axios.delete(
-      `http://127.0.0.1:8000/api/collections/${collection}/cards/${CardId}/`
+      `http://127.0.0.1:8000/api/collections/${activeCollection}/cards/${displayCard.id}/`
     );
     if (response.status === 204) {
       await getCardsforCollection();
@@ -20,31 +24,37 @@ const DeleteCardForm = ({ getCardsforCollection, setDeleteShowModal }) => {
   };
 
   const handleSubmit = (event) => {
-    event.preventDefault();
-    deleteCard();
-    setCollection(0);
-    setCardId("");
-    handleModal();
+    setValidationAnswer(event);
+    if (event === "yes") {
+      deleteCard();
+      handleModal();
+    } else if (event === "no") {
+      handleModal();
+    }
   };
 
   return (
-    <form className="form" onSubmit={handleSubmit}>
+    <div className="form" onSubmit={handleSubmit}>
       <div className="input_group">
+        <div>Are you sure you want to delete this card?</div>
         <label>Collection:</label>
-        <input
-          value={collection}
-          onChange={(event) => setCollection(event.target.value)}
-        />
+        <button
+          value={"yes"}
+          onClick={(event) => handleSubmit(event.target.value)}
+        >
+          Yes
+        </button>
       </div>
       <div className="input_group">
         <label>Card Id:</label>
-        <input
-          value={CardId}
-          onChange={(event) => setCardId(event.target.value)}
-        />
+        <button
+          value={"no"}
+          onClick={(event) => handleSubmit(event.target.value)}
+        >
+          No
+        </button>
       </div>
-      <button type="submit">Delete Card</button>
-    </form>
+    </div>
   );
 };
